modularize into hooks where possible
