+++ apps/coordinator/src/components/Wallet/TransactionsTab/FeeBumping/utils.ts
@@ -54,7 +54,7 @@ export const formatFee = (
  *
  * @internal This is a private validation function used internally
  */
-const validateTransactionInputs = (
+export const validateTransactionInputs = (
   txHex: string,
   fee: number,
   availableUtxos: FeeUTXO[],
@@ -70,130 +70,6 @@ const validateTransactionInputs = (
   if (!availableUtxos?.length) {
     throw new Error("No UTXOs available for fee bumping");
   }
-};
-
-/**
- * Identifies the change output in a transaction by analyzing output addresses
- * and wallet data
- *
- * This function uses multiple heuristics to identify which output is the change:
- * 1. Matches against known wallet addresses
- * 2. Checks BIP32 path patterns (change addresses use path m/1/*)
- * 3. Position in outputs (change is often the last output)
- *
- * @param transaction - The transaction object
- * @param walletState - The wallet state containing addresses
- * @returns Index of the change output or undefined if not found
- *
- * @see https://en.bitcoin.it/wiki/Privacy#Change_address_detection
- */
-export const getChangeOutputIndex = (
-  transaction: TransactionDetails,
-  walletAddresses: string[],
-  changeAddresses: string[],
-): number | undefined => {
-  if (!transaction.vout?.length) return undefined;
-
-  const changeAddressesSet = new Set(changeAddresses);
-  const walletAddressesSet = new Set(walletAddresses);
-
-  // 1) First look for any explicit change‑address hits
-  for (let i = 0; i < transaction.vout.length; i++) {
-    const addr = transaction.vout[i].scriptPubkeyAddress;
-    if (addr && changeAddressesSet.has(addr)) {
-      return i;
-    }
-  }
-
-  // 2) : Check if any output goes to a known wallet address
-  // This is less reliable but can help identify change when the exact
-  // change address isn't recognized
-  for (let i = 0; i < transaction.vout.length; i++) {
-    const addr = transaction.vout[i].scriptPubkeyAddress;
-    if (addr && walletAddressesSet.has(addr)) {
-      return i;
-    }
-  }
-
-  return undefined;
-};
-
-// =============================================================================
-// MAIN ANALYSIS UTIL FUNCTIONS FOR FEE BUMPING
-// =============================================================================
-
-/**
- * Performs comprehensive transaction analysis for fee bumping recommendations
- *
- * This function is the core of the fee bumping system. It uses the Caravan fees
- * package's TransactionAnalyzer to evaluate whether a transaction can be fee-bumped
- * using Replace-by-Fee (RBF) or Child-Pays-for-Parent (CPFP) methods.
- *
- * The analysis considers:
- * - Current transaction fee rate vs. target fee rate
- * - Available UTXOs for additional inputs
- * - Wallet configuration (multisig setup)
- * - Network conditions and fee estimates
- *
- * @param txHex - Raw transaction hex string to analyze
- * @param fee - Current absolute fee amount in satoshis
- * @param network - Bitcoin network (mainnet, testnet, regtest)
- * @param availableUtxos - UTXOs available for fee bumping (from pending tx + wallet)
- * @param targetFeeRate - Desired fee rate in sat/vB based on user priority
- * @param walletConfig - Multisignature wallet configuration parameters
- * @param feePriority - User-selected priority level for reference
- * @returns Comprehensive fee bump recommendation with multiple strategies
- *
- * @throws {Error} When transaction validation fails or insufficient data provided
- *
- * @example
- * const recommendation = analyzeTransaction(
- *   "020000000001...", // transaction hex
- *   5000, // current fee in sats
- *   Network.MAINNET,
- *   [...availableUtxos],
- *   25, // target fee rate
- *   { requiredSigners: 2, totalSigners: 3, addressType: "P2SH" },
- *   FeePriority.MEDIUM
- * );
- */
-export const analyzeTransaction = (
-  txHex: string,
-  fee: number,
-  network: Network,
-  availableUtxos: FeeUTXO[],
-  targetFeeRate: number,
-  walletConfig: {
-    requiredSigners: number;
-    totalSigners: number;
-    addressType: string;
-  },
-  feePriority: FeePriority = FeePriority.MEDIUM,
-): FeeBumpRecommendation => {
-  // Validate inputs
-  validateTransactionInputs(txHex, fee, availableUtxos);
-
-  // Create analyzer with wallet-specific parameters
-  const analyzer = new TransactionAnalyzer({
-    txHex,
-    network,
-    targetFeeRate,
-    absoluteFee: fee.toString(),
-    availableUtxos,
-    requiredSigners: walletConfig.requiredSigners,
-    totalSigners: walletConfig.totalSigners,
-    addressType: walletConfig.addressType as MultisigAddressType,
-  });
-
-  // Get comprehensive analysis
-  const analysis = analyzer.analyze();
-
-  // Return the analysis with added network fee estimates
-  return {
-    ...analysis,
-    userSelectedFeeRate: targetFeeRate,
-    userSelectedPriority: feePriority,
-  };
 };
 
 /**
